Change: Move useCapabilities hook to hooks directory
