+++ src/web/hooks/useCapabilities.js
