+++ src/web/pages/results/detailspage.jsx
@@ -69,7 +69,7 @@ import {getUsername} from 'web/store/usersettings/selectors';
 import compose from 'web/utils/compose';
 import {generateFilename} from 'web/utils/render';
 import PropTypes from 'web/utils/proptypes';
-import useCapabilities from 'web/utils/useCapabilities';
+import useCapabilities from 'web/hooks/useCapabilities';
 
 import NoteComponent from '../notes/component';
